Semi-functional skeleton code
The code can now load and train a basic CNN on the data. It does hwoever crash when it tries to load the .DS_Store file that lies in every dataset folder.

Had to make a custom class that loads the image files. Unsure how relatively fast it is. Might need to look into optimizing it.